xcode 6 and ios 8 issues: a) /Users/saswata_basu/code/oneFrame/oneFrame/MainStoryboard.storyboard:4: Prototype table cells must have reuse identifiers  - fix it by setting the prototype cells to 0 in attribute inspector b) NSInteger fix with either setting type (int) or (unsigned long) c) MKStoreKit shareKit fixes - settings, design, appdelegate d) /Users/saswata_basu/code/oneFrame/oneFrame/MainStoryboard.storyboard:4: Prototype table cells must have reuse identifiers e) oneframe imageview asset errors for collection view - could be due int and/or unsigned long setup f) reverting by checking out git from bitbucket g) GPUImage - add to project, then as target dependency, and then need to copy lib file to resources h) delete library and framework header paths and set to local one i) save to local and then upload mkstore, flurry, and gpuimage to project so that they are all local j) active architecture set to YES to avoid gpuimage linker error k)popupQuery actionSheet set title to nil l) NSInteger into (int) or (unsigned int) m) screenshots n) version number o) what’s new, description, app icon large jpeg format p) ad page changes in storyboard and change link setting q) fill all iphone app icons except car icon; launch image use all iphone images except for iphone 6 and 6+ (HD) r) test on device

+++ oneFrame/pageViewController.m
@@ -58,7 +58,7 @@
     if (([self.pageImages count] == 0) || (index >= [self.pageImages count])) {
         return nil;
     }
-    NSString *string = [NSString stringWithFormat:@"%u of %lu",index+1,(unsigned long)self.pageImages.count];
+    NSString *string = [NSString stringWithFormat:@"%lu of %lu",(unsigned long)index+1,(unsigned long)self.pageImages.count];
     ALAsset *asset = self.pageImages[index];
     ALAssetRepresentation *defaultRep = [asset defaultRepresentation];
     UIImage *image = [UIImage imageWithCGImage:[defaultRep fullScreenImage] scale:[defaultRep scale] orientation:0];
